refactor to clarify relational vs non-relational rules
