+++ respark/rules/generation_rules/generate_datetime.py
@@ -1,6 +1,7 @@
 from datetime import datetime
 from pyspark.sql import Column, functions as F
-from ..rules_registry import register_generation_rule, GenerationRule
+from ..rule_types import GenerationRule
+from ..registry import register_generation_rule
 from respark.random import randint_int, randint_long
