+++ tests/test_rules/test_generation_rules/test_generation_rules.py
@@ -1,7 +1,7 @@
 import pytest
 from datetime import date
 from pyspark.sql import functions as F, types as T
-from respark.rules import get_generation_rule, GENERATION_RULES_REGISTRY
+from respark.rules.registry import get_generation_rule, GENERATION_RULES_REGISTRY
 from respark.core import INTEGRAL_BOUNDS
 
 TEST_SEED = 2025
